post and pull vote auth and resources

+++ app/Http/Controllers/Api/PollVoteController.php
@@ -5,9 +5,16 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Models\PollVote;
 use App\Traits\ResponseJson;
+use App\Traits\MediaTraits;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Models\Permission;
+use Spatie\Permission\Models\Role;
+use Spatie\Permission\PermissionRegistrar;
+use App\Exceptions\NotAuthorized;
+use App\Exceptions\NotFound;
+use App\Http\Resources\PollVoteResource;
 
 class PollVoteController extends Controller
 {
@@ -15,12 +22,17 @@ class PollVoteController extends Controller
    
     public function index()
     {
-        $votes = PollVote::all();
-        if($votes){
-            return $this->jsonResponseWithoutMessage($votes, 'data',200);
+        if(Auth::user()->can('list poll-votes')){
+            $votes = PollVote::all();
+            if($votes){
+                return $this->jsonResponseWithoutMessage(PollVoteResource::collection($votes), 'data',200);
+            }
+            else{
+                throw new NotFound;
+            }
         }
         else{
-           // throw new NotFound;
+            throw new NotAuthorized;
         }
     }
 
@@ -31,16 +43,24 @@ class PollVoteController extends Controller
             'post_id' => 'required',
             'option' => 'required',
         ]);
-     
         if ($validator->fails()) {
             return $this->jsonResponseWithoutMessage($validator->errors(), 'data', 500);
-        }    
+        }   
+
         if(Auth::user()->can('create vote')){
+
+            $option = $request->option;
+            if (is_array($option)){
+                serialize($option);
+            }
+
             PollVote::create($request->all());
+            PollVote::create(new PollVoteResource($request->all()));
+
             return $this->jsonResponseWithoutMessage("Vote Craeted Successfully", 'data', 200);
         }
         else{
-            //throw new NotAuthorized;   
+            throw new NotAuthorized;   
         }
     }
 
@@ -56,10 +76,10 @@ class PollVoteController extends Controller
 
         $vote = PollVote::find($request->poll_vote_id);
         if($vote){
-            return $this->jsonResponseWithoutMessage($vote, 'data',200);
+            return $this->jsonResponseWithoutMessage(new PollVoteResource($vote), 'data',200);
         }
         else{
-           // throw new NotFound;
+           throw new NotFound;
         }
     }
 
@@ -77,11 +97,16 @@ class PollVoteController extends Controller
         }
         if(Auth::user()->can('edit vote')){
             $vote = PollVote::find($request->poll_vote_id);
-            $vote->update($request->all());
-            return $this->jsonResponseWithoutMessage("Vote Updated Successfully", 'data', 200);
+            if ($vote){
+                $vote->update($request->all());
+                return $this->jsonResponseWithoutMessage("Vote Updated Successfully", 'data', 200);
+            }
+            else{
+                throw new NotFound;   
+            }
         }
         else{
-            //throw new NotAuthorized;   
+            throw new NotAuthorized;   
         }
     }
 
@@ -97,8 +122,13 @@ class PollVoteController extends Controller
 
         if(Auth::user()->can('delete vote')){
             $vote = PollVote::find($request->poll_vote_id);
-            $vote->delete();
-            return $this->jsonResponseWithoutMessage("Vote Deleted Successfully", 'data', 200);
+            if($vote){
+                $vote->delete();
+                return $this->jsonResponseWithoutMessage("Vote Deleted Successfully", 'data', 200);
+            }
+            else{
+                throw new NotFound;
+            }
         }
         else{
             //throw new NotAuthorized;
